refactor(tools): split pure layout checks into layout_checks.{h,cpp}
Separate the Max-API-independent layout-check logic from the SDK glue in
layout_tools, making the pure/glue boundary explicit at the file level and
aligning the test layers.

- Move LayoutObject / LayoutCord / LayoutCheckOptions, the append_*_findings
  helpers, and run_layout_checks into src/tools/layout_checks.{h,cpp}
  (depends only on utils/geometry.h + nlohmann/json).
- layout_tools.{h,cpp} now include layout_checks.h and keep only the SDK glue
  (extract_layout_objects/cords, deferred callback, execute_validate_layout,
  schema, dispatch).
- CMake: add layout_checks.cpp to TOOLS_SRC.
- tests: move test_layout_validate.cpp to the pure maxmcp_tests target and
  link layout_checks.cpp into both test targets; switch its include to
  tools/layout_checks.h.

Behavior and API unchanged. ./build.sh --test passes (168 tests); the 9
RunLayoutChecks.* tests now run under the pure maxmcp_tests target. Verified
on-device against the full 27-tool integration checklist.

Closes #80
